% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histogram}
\alias{histogram}
\title{Create a histogram of a single column of data}
\usage{
histogram(
  data,
  xlab = "Values",
  ylab = "Frequency",
  main = "Histogram of Data"
)
}
\arguments{
\item{data}{A vector of numeric data}

\item{xlab}{The label for the x-axis (default: "Values")}

\item{ylab}{The label for the y-axis (default: "Frequency")}

\item{main}{The title of the plot (default: "Histogram of Data")}
}
\description{
This function takes in a single column of data and produces a histogram of its distribution
}
\examples{
data <- rnorm(100)
histogram(data)
}
