% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Perform a t-test on two vectors}
\usage{
t_test(x, y)
}
\arguments{
\item{x}{a vector of numerical values}

\item{y}{a vector of numerical values}
}
\value{
a list containing the t-statistic, degrees of freedom, and p-value for the t-test
}
\description{
This function takes two vectors of numerical values as input and performs a two-sample t-test.
}
\examples{
x <- c(1, 2, 3)
y <- c(4, 5, 6)
t_test(x, y)
$t
[1] -3.464102

$df
[1] 4

$p.value
[1] 0.0194923
}
