% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Time series plot}
\usage{
time_series(
  x,
  y = NULL,
  xlab = "Time",
  ylab = "Value",
  main = "Time Series Plot",
  type = "l",
  col = "blue"
)
}
\arguments{
\item{x}{A vector or data frame containing the time series data}

\item{y}{(optional) A vector or data frame containing additional time series data to plot on the same axis}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{main}{The main plot title}

\item{type}{The type of plot to create (e.g. "l" for a line plot, "p" for a point plot)}

\item{col}{The color(s) to use for the plot(s)}
}
\description{
Creates a time series plot for a given vector or data frame
}
\examples{
time_series(ldeaths)
time_series(ldeaths, type = "p", col = "red")

}
